requirements.txt
Flask
redis
celery
minio
Flask-Migrate
Flask-SQLAlchemy
contextlib2
requests
python-dotenv
pymysql
cryptography